refactor: move several modules to submodules
