+++ dpdispatcher/contexts/dp_cloud_server_context.py
@@ -1,29 +1,32 @@
+#!/usr/bin/env python
+# %%
 import os
 import shutil
 import uuid
+from typing import List
 
 import tqdm
-
-try:
-    from bohriumsdk.client import Client
-    from bohriumsdk.job import Job
-    from bohriumsdk.storage import Storage
-    from bohriumsdk.util import Util
-except ModuleNotFoundError:
-    found_bohriumsdk = False
-else:
-    found_bohriumsdk = True
+from dargs.dargs import Argument
 
 from dpdispatcher import dlog
 from dpdispatcher.base_context import BaseContext
-from dpdispatcher.JobStatus import JobStatus
+from dpdispatcher.dpcloudserver.config import ALI_STS_BUCKET_NAME, ALI_STS_ENDPOINT
+
+# from dpdispatcher.submission import Machine
+# from . import dlog
+from dpdispatcher.utils.dpcloudserver.client import Client
+from dpdispatcher.utils.dpcloudserver.zip_file import zip_file
+
+# from zip_file import zip_files
 
 DP_CLOUD_SERVER_HOME_DIR = os.path.join(
     os.path.expanduser("~"), ".dpdispatcher/", "dp_cloud_server/"
 )
 
 
-class OpenAPIContext(BaseContext):
+class BohriumContext(BaseContext):
+    alias = ("DpCloudServerContext", "LebesgueContext")
+
     def __init__(
         self,
         local_root,
@@ -32,32 +35,47 @@ class OpenAPIContext(BaseContext):
         *args,
         **kwargs,
     ):
-        if not found_bohriumsdk:
-            raise ModuleNotFoundError(
-                "bohriumsdk not installed. Install dpdispatcher with `pip install dpdispatcher[bohrium]`"
-            )
         self.init_local_root = local_root
         self.init_remote_root = remote_root
         self.temp_local_root = os.path.abspath(local_root)
         self.remote_profile = remote_profile
-        self.client = Client()
-        self.storage = Storage(client=self.client)
-        self.job = Job(client=self.client)
-        self.util = Util()
-        self.jgid = None
+        ticket = os.environ.get("BOHR_TICKET", None)
+        email = remote_profile.get("email", None)
+        phone = remote_profile.get("phone", None)
+        password = remote_profile.get("password")
+        os.makedirs(DP_CLOUD_SERVER_HOME_DIR, exist_ok=True)
+
+        if ticket is not None:
+            self.api = Client(ticket=ticket)
+            return
+
+        if email is None and phone is None:
+            raise ValueError(
+                "can not find email/phone number in remote_profile, please check your machine file."
+            )
+        if password is None:
+            raise ValueError(
+                "can not find password in remote_profile, please check your machine file."
+            )
+        # account 作为登录账号
+        account = email
+        if email is None:
+            account = phone
+
+        self.api = Client(account, password)
 
     @classmethod
     def load_from_dict(cls, context_dict):
-        local_root = context_dict.get("local_root", "./")
+        local_root = context_dict["local_root"]
         remote_root = context_dict.get("remote_root", None)
         remote_profile = context_dict.get("remote_profile", {})
 
-        bohrium_context = cls(
+        dp_cloud_server_context = cls(
             local_root=local_root,
             remote_root=remote_root,
             remote_profile=remote_profile,
         )
-        return bohrium_context
+        return dp_cloud_server_context
 
     def bind_submission(self, submission):
         self.submission = submission
@@ -68,57 +86,50 @@ class OpenAPIContext(BaseContext):
 
         self.machine = submission.machine
 
-    def _gen_object_key(self, job, zip_filename):
+        # def zip_files(self, submission):
+        #     file_uuid = uuid.uuid1().hex
+        # oss_task_dir = os.path.join()
+
+    def _gen_oss_path(self, job, zip_filename):
         if hasattr(job, "upload_path") and job.upload_path:
             return job.upload_path
         else:
-            project_id = self.remote_profile.get("project_id")
-
+            program_id = self.remote_profile.get("program_id")
+            program_id = self.remote_profile.get("project_id", program_id)
+            if program_id is None:
+                program_id = 0
             uid = uuid.uuid4()
-            path = os.path.join(str(project_id), str(uid), zip_filename)
+            path = os.path.join("program", str(program_id), str(uid), zip_filename)
             setattr(job, "upload_path", path)
             return path
 
     def upload_job(self, job, common_files=None):
+        MAX_RETRY = 3
         if common_files is None:
             common_files = []
         self.machine.gen_local_script(job)
         zip_filename = job.job_hash + ".zip"
+        oss_task_zip = self._gen_oss_path(job, zip_filename)
         zip_task_file = os.path.join(self.local_root, zip_filename)
 
         upload_file_list = [
             job.script_file_name,
             f"{job.script_file_name}.run",
         ]
-
         upload_file_list.extend(common_files)
 
         for task in job.job_task_list:
             for file in task.forward_files:
                 upload_file_list.append(os.path.join(task.task_work_path, file))
 
-        upload_zip = Util.zip_file_list(
+        upload_zip = zip_file.zip_file_list(
             self.local_root, zip_task_file, file_list=upload_file_list
         )
-        project_id = self.remote_profile.get("project_id", 0)
-
-        data = self.job.create(
-            project_id=project_id,
-            name=self.remote_profile.get("job_name", "DP-GEN"),
-            group_id=self.jgid,  # type: ignore
+        result = self.api.upload(
+            oss_task_zip, upload_zip, ALI_STS_ENDPOINT, ALI_STS_BUCKET_NAME
         )
-        self.jgid = data.get("jobGroupId", "")  # type: ignore
-        token = data.get("token", "")  # type: ignore
-
-        object_key = os.path.join(data["storePath"], zip_filename)  # type: ignore
-        job.upload_path = object_key
-        job.job_id = data["jobId"]  # type: ignore
-        job.jgid = data["jobGroupId"]  # type: ignore
-        self.storage.upload_From_file_multi_part(
-            object_key=object_key, file_path=upload_zip, token=token
-        )
-
-        # self._backup(self.local_root, upload_zip)
+        retry_count = 0
+        self._backup(self.local_root, upload_zip)
 
     def upload(self, submission):
         # oss_task_dir = os.path.join('%s/%s/%s.zip' % ('indicate', file_uuid, file_uuid))
@@ -132,14 +143,13 @@ class OpenAPIContext(BaseContext):
         result = None
         dlog.info("checking all job has been uploaded")
         for job in submission.belonging_jobs:
-            if job.job_state in (JobStatus.unsubmitted, JobStatus.terminated):
-                job_to_be_uploaded.append(job)
+            job_to_be_uploaded.append(job)
         if len(job_to_be_uploaded) == 0:
             dlog.info("all job has been uploaded, continue")
             return result
         for job in tqdm.tqdm(
             job_to_be_uploaded,
-            desc="Uploading to tiefblue",
+            desc="Uploading to Lebesgue",
             bar_format=bar_format,
             leave=False,
             disable=None,
@@ -155,10 +165,10 @@ class OpenAPIContext(BaseContext):
         job_infos = {}
         job_result = []
         for job in jobs:
-            jid = job.job_id
+            ids = job.job_id.split(":job_group_id:")
+            jid = int(ids[0])
             job_hashs[jid] = job.job_hash
-            jobinfo = self.job.detail(jid)
-            # jobinfo = self.api.get_job_detail(jid)
+            jobinfo = self.api.get_job_detail(jid)
             job_result.append(jobinfo)
         # if group_id is not None:
         #     job_result = self.api.get_tasks_list(group_id)
@@ -188,13 +198,38 @@ class OpenAPIContext(BaseContext):
                 target_result_zip, self.local_root
             ):
                 continue
-            self.storage.download_from_url(info["resultUrl"], target_result_zip)
-            Util.unzip_file(target_result_zip, out_dir=self.local_root)
+            self.api.download_from_url(info["resultUrl"], target_result_zip)
+            zip_file.unzip_file(target_result_zip, out_dir=self.local_root)
             self._backup(self.local_root, target_result_zip)
         self._clean_backup(
             self.local_root, keep_backup=self.remote_profile.get("keep_backup", True)
         )
         return True
+
+    def _check_if_job_has_already_downloaded(self, target, local_root):
+        backup_file_location = os.path.join(
+            local_root, "backup", os.path.split(target)[1]
+        )
+        if os.path.exists(backup_file_location):
+            return True
+        else:
+            return False
+
+    def _backup(self, local_root, target):
+        try:
+            # move to backup directory
+            os.makedirs(os.path.join(local_root, "backup"), exist_ok=True)
+            shutil.move(
+                target, os.path.join(local_root, "backup", os.path.split(target)[1])
+            )
+        except (OSError, shutil.Error) as e:
+            dlog.exception("unable to backup file, " + str(e))
+
+    def _clean_backup(self, local_root, keep_backup=True):
+        if not keep_backup:
+            dir_to_be_removed = os.path.join(local_root, "backup")
+            if os.path.exists(dir_to_be_removed):
+                shutil.rmtree(dir_to_be_removed)
 
     def write_file(self, fname, write_str):
         result = self.write_home_file(fname, write_str)
@@ -234,27 +269,72 @@ class OpenAPIContext(BaseContext):
         os.remove(submission_json)
         return True
 
-    def _check_if_job_has_already_downloaded(self, target, local_root):
-        backup_file_location = os.path.join(
-            local_root, "backup", os.path.split(target)[1]
+    # def get_return(self, cmd_pipes):
+    #     if not self.check_finish(cmd_pipes):
+    #         return None, None, None
+    #     else :
+    #         retcode = cmd_pipes['stdout'].channel.recv_exit_status()
+    #         return retcode, cmd_pipes['stdout'], cmd_pipes['stderr']
+
+    @classmethod
+    def machine_subfields(cls) -> List[Argument]:
+        """Generate the machine subfields.
+
+        Returns
+        -------
+        list[Argument]
+            machine subfields
+        """
+        doc_remote_profile = (
+            "The information used to maintain the connection with remote machine."
         )
-        if os.path.exists(backup_file_location):
-            return True
-        else:
-            return False
-
-    def _backup(self, local_root, target):
-        try:
-            # move to backup directory
-            os.makedirs(os.path.join(local_root, "backup"), exist_ok=True)
-            shutil.move(
-                target, os.path.join(local_root, "backup", os.path.split(target)[1])
+        doc_retry_count = "The retry count when a job is terminated"
+        doc_ignore_exit_code = """The job state will be marked as finished if the exit code is non-zero when set to True. Otherwise,
+              the job state will be designated as terminated."""
+        return [
+            Argument(
+                "remote_profile",
+                dict,
+                [
+                    Argument("email", str, optional=True, doc="Email"),
+                    Argument("password", str, optional=True, doc="Password"),
+                    Argument(
+                        "program_id",
+                        int,
+                        optional=False,
+                        alias=["project_id"],
+                        doc="Program ID",
+                    ),
+                    Argument(
+                        "retry_count",
+                        [int, type(None)],
+                        optional=True,
+                        default=2,
+                        doc=doc_retry_count,
+                    ),
+                    Argument(
+                        "ignore_exit_code",
+                        bool,
+                        optional=True,
+                        default=True,
+                        doc=doc_ignore_exit_code,
+                    ),
+                    Argument(
+                        "keep_backup",
+                        bool,
+                        optional=True,
+                        doc="keep download and upload zip",
+                    ),
+                    Argument(
+                        "input_data", dict, optional=False, doc="Configuration of job"
+                    ),
+                ],
+                doc=doc_remote_profile,
             )
-        except (OSError, shutil.Error) as e:
-            dlog.exception("unable to backup file, " + str(e))
+        ]
 
-    def _clean_backup(self, local_root, keep_backup=True):
-        if not keep_backup:
-            dir_to_be_removed = os.path.join(local_root, "backup")
-            if os.path.exists(dir_to_be_removed):
-                shutil.rmtree(dir_to_be_removed)
+
+DpCloudServerContext = BohriumContext
+LebesgueContext = BohriumContext
+
+# %%
