+++ dpdispatcher/contexts/lazy_local_context.py
