+++ dpdispatcher/contexts/ssh_context.py
@@ -22,7 +22,7 @@ from dpdispatcher import dlog
 from dpdispatcher.base_context import BaseContext
 
 # from dpdispatcher.submission import Machine
-from dpdispatcher.utils import RetrySignal, generate_totp, get_sha256, retry, rsync
+from dpdispatcher.utils.utils import RetrySignal, generate_totp, get_sha256, retry, rsync
 
 
 class SSHSession:
