+++ dpdispatcher/entrypoints/__init__.py
@@ -1,0 +1,1 @@
+"""Entry points."""
