+++ dpdispatcher/machines/dp_cloud_server.py
@@ -7,9 +7,9 @@ import warnings
 from dpdispatcher import dlog
 from dpdispatcher.dpcloudserver import Client, zip_file
 from dpdispatcher.dpcloudserver.config import ALI_OSS_BUCKET_URL
-from dpdispatcher.JobStatus import JobStatus
+from dpdispatcher.utils.JobStatus import JobStatus
 from dpdispatcher.machine import Machine
-from dpdispatcher.utils import customized_script_header_template
+from dpdispatcher.utils.utils import customized_script_header_template
 
 shell_script_header_template = """
 #!/bin/bash -l
