+++ dpdispatcher/machines/shell.py
@@ -1,9 +1,9 @@
 import shlex
 
 from dpdispatcher import dlog
-from dpdispatcher.JobStatus import JobStatus
+from dpdispatcher.utils.JobStatus import JobStatus
 from dpdispatcher.machine import Machine
-from dpdispatcher.utils import customized_script_header_template
+from dpdispatcher.utils.utils import customized_script_header_template
 
 shell_script_header_template = """
 #!/bin/bash -l
