+++ dpdispatcher/machines/slurm.py
@@ -6,9 +6,9 @@ from typing import List
 from dargs import Argument
 
 from dpdispatcher import dlog
-from dpdispatcher.JobStatus import JobStatus
+from dpdispatcher.utils.JobStatus import JobStatus
 from dpdispatcher.machine import Machine, script_command_template
-from dpdispatcher.utils import RetrySignal, customized_script_header_template, retry
+from dpdispatcher.utils.utils import RetrySignal, customized_script_header_template, retry
 
 # from dpdispatcher.submission import Resources
